Refactor ANN SIMD funcs
